Add package info to integration
