+++ packages/babel-plugin/package.json
@@ -22,6 +22,6 @@
   "dependencies": {
     "@babel/core": "^7.13.10",
     "@babel/template": "^7.12.13",
-    "find-up": "^5.0.0"
+    "@vanilla-extract/integration": "^0.1.0"
   }
 }
