+++ packages/integration/src/filters.ts
@@ -1,0 +1,2 @@
+export const cssFileFilter = /\.css\.(js|jsx|ts|tsx)$/;
+export const virtualCssFileFilter = /vanilla\.css\?source=.*$/;
