+++ packages/integration/src/index.ts
@@ -1,5 +1,4 @@
 export { processVanillaFile } from './processVanillaFile';
 export { getSourceFromVirtualCssFile } from './virtualFile';
-
-export const cssFileFilter = /\.css\.(js|jsx|ts|tsx)$/;
-export const virtualCssFileFilter = /vanilla\.css\?source=.*$/;
+export { getPackageInfo } from './packageInfo';
+export * from './filters';
