+++ packages/webpack-plugin/package.json
@@ -31,19 +31,13 @@
   "dependencies": {
     "@vanilla-extract/css": "^0.4.0",
     "@vanilla-extract/integration": "^0.1.0",
-    "chalk": "^4.1.0",
     "debug": "^4.3.1",
-    "dedent": "^0.7.0",
-    "eval": "^0.1.6",
-    "javascript-stringify": "^2.0.1",
     "loader-utils": "^2.0.0",
     "lodash": "^4.17.21",
     "virtual-resource-loader": "^1.0.0"
   },
   "devDependencies": {
     "@types/debug": "^4.1.5",
-    "@types/dedent": "^0.7.0",
-    "@types/lodash": "^4.14.168",
     "webpack": "^5.27.1"
   }
 }
